Decluttered CSS. Added New-Game button with logic, Added stats menu with very basic logic
added new game button and functionality. changed wordle.js to return an object of functions so that the global variables dont leak out to other js files

added some functionality for displaying stats

remade stats menu which also gives the user a message upon game over

Attemmpt to adjust mobile styling issues. 1) When zoomed in, the page jerks the page left upon an input, so I made the hidden input area readonly in the hopes that it can still be focused. 2) Added a media min-width: 600 so the board appears larger on the users screen